requirements.txt
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.2
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
edgartools==4.18.0
filelock==3.20.0
google-auth==2.41.1
google-genai==1.43.0
h11==0.16.0
hishel==0.1.3
httpcore==1.0.9
httpx==0.28.1
httpxthrottlecache==0.2.1
humanize==4.13.0
idna==3.10
jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.2.43
langgraph==0.1.18
langsmith==0.1.147
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==2.3.3
ollama==0.6.0
orjson==3.11.3
packaging==24.2
pandas==2.2.2
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.0
pydantic-core==2.41.1
pygments==2.19.2
pygraphviz==1.13
pyrate-limiter==3.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.3
rank-bm25==0.2.2
rapidfuzz==3.14.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stamina==25.1.0
tabulate==0.9.0
tenacity==8.5.0
textdistance==4.6.3
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
unidecode==1.4.0
urllib3==2.5.0
websockets==15.0.1
